requirements.txt
python-binance
TA-Lib
csv